Extract filter types/constants to types/filters
