+++ frontend/src/components/FilterSidebar.tsx
@@ -1,30 +1,10 @@
 import { Minus, Plus, SlidersHorizontal, X } from "lucide-react";
-
-export const COMMON_AMENITIES = [
-    "WiFi", "Parking", "Pet Friendly", "TV", "AC", "Kitchen", "Pool", "Gym", "Breakfast",
-];
-
-export const PROPERTY_TYPES = ["Apartment", "House", "Villa", "Studio", "Room"];
-
-export interface FilterState {
-    selectedAmenities: string[];
-    minPrice: string;
-    maxPrice: string;
-    propertyTypes: string[];
-    guests: number;
-    bedrooms: number;
-    beds: number;
-}
-
-export const EMPTY_FILTERS: FilterState = {
-    selectedAmenities: [],
-    minPrice: "",
-    maxPrice: "",
-    propertyTypes: [],
-    guests: 1,
-    bedrooms: 0,
-    beds: 0,
-};
+import {
+    COMMON_AMENITIES,
+    EMPTY_FILTERS,
+    PROPERTY_TYPES,
+    type FilterState,
+} from "../types/filters";
 
 interface FilterSidebarProps {
     filters: FilterState;
